+++ R/cosewic_tools.R
@@ -115,7 +115,7 @@ cosewic_ranges <- function(df_db,
                            iao_grid_size_km = 2,
                            eoo_p = 0.95,
                            eoo_clip = NULL,
-                           crs = crs_albers_canada(),
+                           crs = "ESRI:102001",
                            filter_unique = FALSE,
                            spatial = TRUE) {
   
@@ -370,7 +370,7 @@ prep_spatial <- function(df,
 #'   labs(caption = "No buffer")
  
 grid_canada <- function(cell_size = 200, buffer = 500, 
-                        crs = crs_albers_canada()){
+                        crs = "ESRI:102001"){
   have_pkg_check("sf")
   map_canada(crs = crs) %>%
     sf::st_buffer(units::set_units(buffer, "km")) %>%
@@ -613,45 +613,3 @@ cosewic_plot_indiv <- function(e, a, points, grid, map, scale, title, crs, verbo
   g
 }
 
-
-#' Create the Canada Albers Equal Area projection
-#' 
-#' For area calculations it's best to use a projection which preserves area. 
-#' The Equal Areas Conic Projection is good for this, and can be customized to
-#' Canada.
-#' 
-#' This definition is from https://epsg.io/102001.
-#'
-#' @returns CRS
-#' @export
-#'
-#' @examples
-#' crs_albers_canada()
-
-crs_albers_canada <- function() {
-  sf::st_crs(
-  'PROJCS["Canada Albers Equal Area Conic",
-    GEOGCS["NAD83",
-        DATUM["North_American_Datum_1983",
-            SPHEROID["GRS 1980",6378137,298.257222101,
-                AUTHORITY["EPSG","7019"]],
-            AUTHORITY["EPSG","6269"]],
-        PRIMEM["Greenwich",0,
-            AUTHORITY["EPSG","8901"]],
-        UNIT["degree",0.0174532925199433,
-            AUTHORITY["EPSG","9122"]],
-        AUTHORITY["EPSG","4269"]],
-    PROJECTION["Albers_Conic_Equal_Area"],
-    PARAMETER["latitude_of_center",40],
-    PARAMETER["longitude_of_center",-96],
-    PARAMETER["standard_parallel_1",50],
-    PARAMETER["standard_parallel_2",70],
-    PARAMETER["false_easting",0],
-    PARAMETER["false_northing",0],
-    UNIT["metre",1,
-        AUTHORITY["EPSG","9001"]],
-    AXIS["Easting",EAST],
-    AXIS["Northing",NORTH],
-    AUTHORITY["ESRI","102001"]]')
-}
-